Finish renaming column_info to data_info
